requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coverage==5.5
coveralls==1.11.1
cryptography==44.0.2
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.4.1
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
flake8==7.1.1
idna==3.10
lxml==5.3.1
mccabe==0.7.0
oauthlib==3.2.2
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
responses==0.12.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
unittest-xml-reporting==3.2.0
urllib3==2.3.0
